Modified models

+++ WeatherApp/WeatherApp/Model/Coordinate.swift
@@ -1,5 +1,5 @@
 import Foundation
 
 struct Coordinate: Codable {
-    let lat, lon: Int
+    let lat, lon: Double
 }
